requirements.txt
PySide6
pyautogui
pupil-labs-realtime-api
git+https://github.com/pupil-labs/real-time-screen-gaze
